refactor: structural modifications
