+++ src/components/core/cards/nft-skeleton/nft-skeleton.tsx
