+++ src/components/core/cards/nft-skeleton/styles.ts
@@ -1,4 +1,4 @@
-import { styled } from '../../../stitches.config';
+import { styled } from '../../../../stitches.config';
 
 export const CardWrapper = styled('div', {
   display: 'flex',
