+++ src/components/core/checkbox/checkbox.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Wrapper } from './styles';
 
 export type CheckboxProps = {
