+++ src/components/core/index.ts
@@ -5,6 +5,5 @@ export * from './chips';
 export * from './accordions';
 export * from './checkbox';
 export * from './table-cells';
-export * from './nft-card';
 export * from './status-indicators';
-export * from './nft-skeleton';
+export * from './cards';
